Reorganize functions. Polish code.
- Polish docstrings.
- Fix flake8 errors.
